+++ src/app/shared/services/loader.service.ts
@@ -1,0 +1,52 @@
+import { Injectable } from '@angular/core';
+export interface Scripts {
+  name: string;
+  src: string;
+}
+
+export const ScriptStore: Scripts[] = [
+  { name: 'map', src: 'http://maps.googleapis.com/maps/api/js?libraries=places&sensor=false&key=AIzaSyAeWGJ5mJmTeYmLOhRu6TQ88p7nS5czlw4' },
+];
+
+@Injectable({
+  providedIn: 'root'
+})
+export class LoaderService {
+
+  public loadAPI: Promise<any>;
+  private scripts: any = {};
+
+  constructor() {
+    ScriptStore.forEach((script: any) => {
+      this.scripts[script.name] = {
+        loaded: false,
+        src: script.src
+      };
+    });
+  }
+
+  load(...scripts: string[]) {
+    const promises: any[] = [];
+    scripts.forEach((script) => promises.push(this.loadScript(script)));
+    return Promise.all(promises);
+  }
+
+  loadScript(name: string) {
+    return new Promise((resolve, reject) => {
+      if (!this.scripts[name].loaded) {
+        //load script
+        let script = document.createElement('script');
+        script.type = 'text/javascript';
+        script.src = this.scripts[name].src;
+        script.onload = () => {
+          this.scripts[name].loaded = true;
+          resolve({ script: name, loaded: true, status: 'Loaded' });
+        };
+        script.onerror = (error: any) => reject({ script: name, loaded: false, status: 'Loaded' });
+        document.getElementsByTagName('head')[0].appendChild(script);
+      } else {
+        resolve({ script: name, loaded: true, status: 'Already Loaded' });
+      }
+    });
+  }
+}
