+++ src/app/shared/services/map.service.ts
@@ -1,42 +1,26 @@
 import { Injectable } from '@angular/core';
 
+import { LoaderService } from './loader.service';
 export interface Map {
   lat: any,
   lng: any
 }
 declare let google: any;
 
-export interface Scripts {
-  name: string;
-  src: string;
-}
-
-export const ScriptStore: Scripts[] = [
-  { name: 'map', src: 'http://maps.googleapis.com/maps/api/js?libraries=places&sensor=false&key=AIzaSyAeWGJ5mJmTeYmLOhRu6TQ88p7nS5czlw4' },
-];
-
 @Injectable({
   providedIn: 'root'
 })
 export class MapService {
-
-  public loadAPI: Promise<any>;
-  private scripts: any = {};
+ 
   mapContent: Map[];
   place: any;
   statusColor: string;
 
-  constructor() {
-    ScriptStore.forEach((script: any) => {
-      this.scripts[script.name] = {
-        loaded: false,
-        src: script.src
-      };
-    });
+  constructor(private loaderService: LoaderService) {
   }
 
   plotLocation(structureArray) {
-    this.load('map').then(res => {
+    this.loaderService.load('map').then(res => {
       var map = new google.maps.Map(document.getElementById('map'), {
         zoom: 10,
         center: new google.maps.LatLng(-33.92, 151.25),
@@ -81,30 +65,7 @@ export class MapService {
     })
   }
 
-  load(...scripts: string[]) {
-    const promises: any[] = [];
-    scripts.forEach((script) => promises.push(this.loadScript(script)));
-    return Promise.all(promises);
-  }
-
-  loadScript(name: string) {
-    return new Promise((resolve, reject) => {
-      if (!this.scripts[name].loaded) {
-        //load script
-        let script = document.createElement('script');
-        script.type = 'text/javascript';
-        script.src = this.scripts[name].src;
-        script.onload = () => {
-          this.scripts[name].loaded = true;
-          resolve({ script: name, loaded: true, status: 'Loaded' });
-        };
-        script.onerror = (error: any) => reject({ script: name, loaded: false, status: 'Loaded' });
-        document.getElementsByTagName('head')[0].appendChild(script);
-      } else {
-        resolve({ script: name, loaded: true, status: 'Already Loaded' });
-      }
-    });
-  }
+ 
 
   setLocation(lat, lng) {
     this.mapContent = [];
@@ -117,7 +78,7 @@ export class MapService {
 
   searchLocation(input) {
     var markers;
-    this.load('map').then(res => {
+    this.loaderService.load('map').then(res => {
       var maps = new google.maps.Map(document.getElementById('mapCase'), {
         center: { lat: -34.397, lng: 150.644 },
         zoom: 8,
